refactor(cli): separate command workflows
