+++ crates/omniinfer-cli/src/serve/mod.rs
@@ -25,495 +25,12 @@ const SHUTDOWN_REQUEST_TIMEOUT: Duration = Duration::from_secs(10);
 const GRACEFUL_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(3);
 const FORCED_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(5);
 
-pub(crate) fn stop_serve(port: u16) -> Result<()> {
-    let _port_lock = serve_state::try_lock_serve_port(port)?;
-    let info = serve_state::load_serve_pid_info(port)?;
-    stop_serve_locked(port, info, true)
-}
+mod lifecycle;
 
-fn stop_serve_locked(
-    port: u16,
-    info: Option<serve_state::ServePidInfo>,
-    print_status: bool,
-) -> Result<()> {
-    if let Some(info) = info.as_ref() {
-        validate_recorded_processes(info, port)?;
-    }
-    let mut config = config::load_app_config().unwrap_or_default();
-    config.port = port;
-    let url = format!("{}/omni/shutdown", config.service_base_url());
-    let shutdown_accepted =
-        match http_client::post_json(&url, &serde_json::json!({}), SHUTDOWN_REQUEST_TIMEOUT) {
-            Ok(response) => response.status < 400,
-            Err(_) => false,
-        };
-
-    let mut gateway_closed = wait_for_local_port_closed(
-        port,
-        if shutdown_accepted {
-            GRACEFUL_SHUTDOWN_TIMEOUT
-        } else {
-            Duration::ZERO
-        },
-    );
-    gateway_closed = gateway_closed
-        && recorded_process_exited(
-            info.as_ref().and_then(|value| value.pid),
-            info.as_ref()
-                .and_then(|value| value.gateway_process.as_ref()),
-            LegacyProcessKind::Gateway,
-            info.as_ref(),
-            port,
-        );
-    let mut backend_closed = info
-        .as_ref()
-        .and_then(|value| value.backend_port)
-        .is_none_or(|backend_port| wait_for_local_port_closed(backend_port, Duration::ZERO));
-    backend_closed = backend_closed
-        && recorded_process_exited(
-            info.as_ref().and_then(|value| value.backend_pid),
-            info.as_ref()
-                .and_then(|value| value.backend_process.as_ref()),
-            LegacyProcessKind::Backend,
-            info.as_ref(),
-            port,
-        );
-    if info.is_some() && (!gateway_closed || !backend_closed) {
-        if let Some(pid) = info.as_ref().and_then(|value| value.backend_pid) {
-            stop_recorded_process(
-                pid,
-                info.as_ref()
-                    .and_then(|value| value.backend_process.as_ref()),
-                LegacyProcessKind::Backend,
-                info.as_ref(),
-                port,
-            )?;
-        }
-        if let Some(pid) = info.as_ref().and_then(|value| value.pid) {
-            stop_recorded_process(
-                pid,
-                info.as_ref()
-                    .and_then(|value| value.gateway_process.as_ref()),
-                LegacyProcessKind::Gateway,
-                info.as_ref(),
-                port,
-            )?;
-        }
-        gateway_closed = wait_for_local_port_closed(port, FORCED_SHUTDOWN_TIMEOUT);
-        gateway_closed = gateway_closed
-            && wait_for_recorded_process_exit(
-                info.as_ref().and_then(|value| value.pid),
-                info.as_ref()
-                    .and_then(|value| value.gateway_process.as_ref()),
-                LegacyProcessKind::Gateway,
-                info.as_ref(),
-                port,
-                FORCED_SHUTDOWN_TIMEOUT,
-            );
-        backend_closed = info
-            .as_ref()
-            .and_then(|value| value.backend_port)
-            .is_none_or(|backend_port| {
-                wait_for_local_port_closed(backend_port, FORCED_SHUTDOWN_TIMEOUT)
-            });
-        backend_closed = backend_closed
-            && wait_for_recorded_process_exit(
-                info.as_ref().and_then(|value| value.backend_pid),
-                info.as_ref()
-                    .and_then(|value| value.backend_process.as_ref()),
-                LegacyProcessKind::Backend,
-                info.as_ref(),
-                port,
-                FORCED_SHUTDOWN_TIMEOUT,
-            );
-    }
-    let mut tunnel_closed = true;
-    if let Some(pid) = info.as_ref().and_then(|value| value.cloudflared_pid) {
-        stop_recorded_process(
-            pid,
-            info.as_ref()
-                .and_then(|value| value.cloudflared_process.as_ref()),
-            LegacyProcessKind::Cloudflared,
-            info.as_ref(),
-            port,
-        )?;
-        tunnel_closed = wait_for_recorded_process_exit(
-            Some(pid),
-            info.as_ref()
-                .and_then(|value| value.cloudflared_process.as_ref()),
-            LegacyProcessKind::Cloudflared,
-            info.as_ref(),
-            port,
-            FORCED_SHUTDOWN_TIMEOUT,
-        );
-    }
-
-    if (shutdown_accepted || info.is_some()) && gateway_closed && backend_closed && tunnel_closed {
-        remove_recorded_serve_state(port, info.as_ref())?;
-        if print_status {
-            println!("OmniInfer service stopped on port {port}");
-        }
-        Ok(())
-    } else if !shutdown_accepted && info.is_none() {
-        if print_status {
-            println!("OmniInfer service is not running on port {port}");
-        }
-        Ok(())
-    } else {
-        anyhow::bail!("failed to stop OmniInfer service on port {port} within the shutdown timeout")
-    }
-}
-
-#[derive(Clone, Copy)]
-enum LegacyProcessKind {
-    Gateway,
-    Cloudflared,
-    Backend,
-}
-
-fn validate_recorded_processes(info: &serve_state::ServePidInfo, port: u16) -> Result<()> {
-    for (pid, identity, kind, label) in [
-        (
-            info.pid,
-            info.gateway_process.as_ref(),
-            LegacyProcessKind::Gateway,
-            "gateway",
-        ),
-        (
-            info.cloudflared_pid,
-            info.cloudflared_process.as_ref(),
-            LegacyProcessKind::Cloudflared,
-            "cloudflared",
-        ),
-        (
-            info.backend_pid,
-            info.backend_process.as_ref(),
-            LegacyProcessKind::Backend,
-            "backend",
-        ),
-    ] {
-        let Some(pid) = pid else {
-            continue;
-        };
-        if let Some(identity) = identity {
-            if identity.pid != pid
-                || serve_state::process_identity_status(identity)
-                    == serve_state::ProcessIdentityStatus::Mismatched
-            {
-                anyhow::bail!(
-                    "refusing to stop {label} PID {pid}: process identity does not match serve state"
-                );
-            }
-        } else if legacy_process_status(pid, kind, info, port) == LegacyProcessStatus::Mismatched {
-            anyhow::bail!(
-                "refusing to stop legacy {label} PID {pid}: process ownership could not be verified"
-            );
-        }
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum LegacyProcessStatus {
-    Running,
-    Exited,
-    Mismatched,
-}
-
-fn legacy_process_status(
-    pid: u32,
-    kind: LegacyProcessKind,
-    info: &serve_state::ServePidInfo,
-    port: u16,
-) -> LegacyProcessStatus {
-    use sysinfo::{ProcessRefreshKind, ProcessesToUpdate, System, UpdateKind};
-
-    let pid = sysinfo::Pid::from_u32(pid);
-    let mut system = System::new();
-    system.refresh_processes_specifics(
-        ProcessesToUpdate::Some(&[pid]),
-        true,
-        ProcessRefreshKind::nothing()
-            .with_cmd(UpdateKind::Always)
-            .with_exe(UpdateKind::Always)
-            .without_tasks(),
-    );
-    let Some(process) = system.process(pid) else {
-        return LegacyProcessStatus::Exited;
-    };
-    let args = process
-        .cmd()
-        .iter()
-        .map(|value| value.to_string_lossy())
-        .collect::<Vec<_>>();
-    let joined = args.join(" ");
-    let joined_lower = joined.to_ascii_lowercase();
-    let port_text = port.to_string();
-    let matches = match kind {
-        LegacyProcessKind::Gateway => {
-            joined_lower.contains("gateway")
-                && args
-                    .windows(2)
-                    .any(|pair| pair[0] == "--port" && pair[1] == port_text)
-        }
-        LegacyProcessKind::Cloudflared => {
-            joined_lower.contains("cloudflared")
-                && joined_lower.contains("tunnel")
-                && joined.contains(&format!("127.0.0.1:{port}"))
-        }
-        LegacyProcessKind::Backend => {
-            let backend_port = info.backend_port.map(|value| value.to_string());
-            let model = info
-                .model
-                .as_deref()
-                .filter(|value| !value.trim().is_empty());
-            backend_port
-                .as_ref()
-                .is_some_and(|value| joined.contains(value))
-                && model.is_some_and(|value| joined.contains(value))
-        }
-    };
-    if matches {
-        LegacyProcessStatus::Running
-    } else {
-        LegacyProcessStatus::Mismatched
-    }
-}
-
-fn recorded_process_exited(
-    pid: Option<u32>,
-    identity: Option<&serve_state::ProcessIdentity>,
-    kind: LegacyProcessKind,
-    info: Option<&serve_state::ServePidInfo>,
-    port: u16,
-) -> bool {
-    let Some(pid) = pid else {
-        return true;
-    };
-    if let Some(identity) = identity {
-        return serve_state::process_identity_status(identity)
-            != serve_state::ProcessIdentityStatus::Running;
-    }
-    info.is_some_and(|info| {
-        legacy_process_status(pid, kind, info, port) != LegacyProcessStatus::Running
-    })
-}
-
-fn wait_for_recorded_process_exit(
-    pid: Option<u32>,
-    identity: Option<&serve_state::ProcessIdentity>,
-    kind: LegacyProcessKind,
-    info: Option<&serve_state::ServePidInfo>,
-    port: u16,
-    timeout: Duration,
-) -> bool {
-    let deadline = Instant::now() + timeout;
-    while !recorded_process_exited(pid, identity, kind, info, port) && Instant::now() < deadline {
-        thread::sleep(Duration::from_millis(50));
-    }
-    recorded_process_exited(pid, identity, kind, info, port)
-}
-
-fn stop_recorded_process(
-    pid: u32,
-    identity: Option<&serve_state::ProcessIdentity>,
-    kind: LegacyProcessKind,
-    info: Option<&serve_state::ServePidInfo>,
-    port: u16,
-) -> Result<()> {
-    if !recorded_process_exited(Some(pid), identity, kind, info, port) {
-        stop_process(pid);
-    }
-    Ok(())
-}
-
-fn remove_recorded_serve_state(port: u16, info: Option<&serve_state::ServePidInfo>) -> Result<()> {
-    if let Some(run_id) = info.and_then(|value| value.run_id.as_deref()) {
-        serve_state::remove_serve_pid_info_if_run_id(port, run_id)?;
-    } else {
-        serve_state::remove_serve_pid_info(port)?;
-    }
-    Ok(())
-}
-
-fn wait_for_local_port_closed(port: u16, timeout: Duration) -> bool {
-    let deadline = Instant::now() + timeout;
-    while Instant::now() < deadline {
-        if TcpStream::connect(("127.0.0.1", port)).is_err() {
-            return true;
-        }
-        thread::sleep(Duration::from_millis(100));
-    }
-    TcpStream::connect(("127.0.0.1", port)).is_err()
-}
-
-fn ensure_serve_port_available(config: &config::AppConfig) -> Result<()> {
-    #[cfg(debug_assertions)]
-    if env_flag("OMNIINFER_TEST_ALLOW_OCCUPIED_SERVE_PORT") {
-        return Ok(());
-    }
-    match TcpListener::bind((config.host.as_str(), config.port)) {
-        Ok(listener) => {
-            drop(listener);
-            Ok(())
-        }
-        Err(error) => anyhow::bail!(
-            "cannot start OmniInfer service: {}:{} is already in use ({error}). Stop the existing service or choose another --port.",
-            config.host,
-            config.port
-        ),
-    }
-}
-
-fn cleanup_failed_serve(
-    gateway: &mut std::process::Child,
-    cloudflared: Option<&mut std::process::Child>,
-    port: u16,
-    run_id: &str,
-) {
-    if let Some(tunnel) = cloudflared {
-        stop_process(tunnel.id());
-        let _ = wait_for_child_exit(tunnel, FORCED_SHUTDOWN_TIMEOUT);
-    }
-
-    let mut config = config::load_app_config().unwrap_or_default();
-    config.port = port;
-    let url = format!("{}/omni/shutdown", config.service_base_url());
-    let shutdown_accepted =
-        http_client::post_json(&url, &serde_json::json!({}), SHUTDOWN_REQUEST_TIMEOUT)
-            .is_ok_and(|response| response.status < 400);
-    let mut gateway_exited = wait_for_child_exit(
-        gateway,
-        if shutdown_accepted {
-            GRACEFUL_SHUTDOWN_TIMEOUT
-        } else {
-            Duration::ZERO
-        },
-    );
-    let mut gateway_closed = wait_for_local_port_closed(
-        port,
-        if shutdown_accepted {
-            GRACEFUL_SHUTDOWN_TIMEOUT
-        } else {
-            Duration::ZERO
-        },
-    );
-    if !shutdown_accepted || !gateway_exited || !gateway_closed {
-        if !gateway_exited {
-            stop_process(gateway.id());
-        }
-        gateway_exited = gateway_exited || wait_for_child_exit(gateway, FORCED_SHUTDOWN_TIMEOUT);
-        gateway_closed =
-            gateway_closed || wait_for_local_port_closed(port, FORCED_SHUTDOWN_TIMEOUT);
-    }
-
-    if gateway_exited && gateway_closed {
-        let _ = serve_state::remove_serve_pid_info_if_run_id(port, run_id);
-    } else {
-        eprintln!(
-            "OmniInfer: failed startup cleanup did not fully stop the gateway on port {port}; run `omniinfer serve stop --port {port}`"
-        );
-    }
-}
-
-fn cleanup_smoke_serve(
-    gateway: &mut std::process::Child,
-    cloudflared: Option<&mut std::process::Child>,
-    port: u16,
-    backend_pid: Option<u32>,
-    backend_port: Option<u16>,
-    run_id: &str,
-) -> Result<()> {
-    let mut cleanup_errors = Vec::new();
-    if let Some(tunnel) = cloudflared
-        && !wait_for_child_exit(tunnel, Duration::ZERO)
-    {
-        stop_process(tunnel.id());
-        if !wait_for_child_exit(tunnel, FORCED_SHUTDOWN_TIMEOUT) {
-            cleanup_errors.push("cloudflared did not exit".to_string());
-        }
-    }
-
-    let mut config = config::load_app_config().unwrap_or_default();
-    config.port = port;
-    let url = format!("{}/omni/shutdown", config.service_base_url());
-    let shutdown_accepted =
-        http_client::post_json(&url, &serde_json::json!({}), SHUTDOWN_REQUEST_TIMEOUT)
-            .is_ok_and(|response| response.status < 400);
-
-    let gateway_exited = wait_for_child_exit(gateway, GRACEFUL_SHUTDOWN_TIMEOUT);
-    let gateway_closed = wait_for_local_port_closed(port, GRACEFUL_SHUTDOWN_TIMEOUT);
-    let backend_closed = backend_port
-        .is_none_or(|value| wait_for_local_port_closed(value, GRACEFUL_SHUTDOWN_TIMEOUT));
-    if !shutdown_accepted || !gateway_exited || !gateway_closed || !backend_closed {
-        if let Some(pid) = backend_pid {
-            stop_process(pid);
-        }
-        if !gateway_exited {
-            stop_process(gateway.id());
-        }
-    }
-
-    if !wait_for_child_exit(gateway, FORCED_SHUTDOWN_TIMEOUT) {
-        cleanup_errors.push("gateway did not exit".to_string());
-    }
-    if !wait_for_local_port_closed(port, FORCED_SHUTDOWN_TIMEOUT) {
-        cleanup_errors.push(format!("gateway port {port} is still in use"));
-    }
-    if let Some(backend_port) = backend_port
-        && !wait_for_local_port_closed(backend_port, FORCED_SHUTDOWN_TIMEOUT)
-    {
-        cleanup_errors.push(format!("backend port {backend_port} is still in use"));
-    }
-    if cleanup_errors.is_empty() {
-        serve_state::remove_serve_pid_info_if_run_id(port, run_id)?;
-        println!("Smoke test cleanup complete; gateway/backend stopped and port {port} released");
-        Ok(())
-    } else {
-        anyhow::bail!(
-            "smoke test cleanup failed: {}; serve metadata was retained for `omniinfer serve stop --port {port}`",
-            cleanup_errors.join("; ")
-        )
-    }
-}
-
-fn wait_for_child_exit(child: &mut std::process::Child, timeout: Duration) -> bool {
-    let deadline = Instant::now() + timeout;
-    loop {
-        match child.try_wait() {
-            Ok(Some(_)) => {
-                let _ = child.wait();
-                return true;
-            }
-            Ok(None) if Instant::now() < deadline => {
-                thread::sleep(Duration::from_millis(50));
-            }
-            Ok(None) | Err(_) => return false,
-        }
-    }
-}
-
-fn stop_process(pid: u32) {
-    #[cfg(unix)]
-    {
-        let _ = ProcessCommand::new("kill")
-            .arg(pid.to_string())
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status();
-    }
-    #[cfg(windows)]
-    {
-        let mut command = ProcessCommand::new("taskkill");
-        command
-            .args(["/PID", &pid.to_string(), "/T", "/F"])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null());
-        hide_child_window(&mut command);
-        let _ = command.status();
-    }
-}
-
+pub(crate) use lifecycle::stop_serve;
+use lifecycle::{
+    cleanup_failed_serve, cleanup_smoke_serve, ensure_serve_port_available, stop_serve_locked,
+};
 pub(crate) fn can_serve_locally(args: &ServeArgs) -> bool {
     args.command.is_none()
 }
@@ -884,303 +401,10 @@ pub(crate) fn serve_orchestrated(args: &ServeArgs) -> Result<()> {
     Ok(())
 }
 
-#[derive(Debug, Clone)]
-struct ServeModelRequest {
-    model: String,
-    mmproj: Option<String>,
-    ctx_size: Option<u32>,
-    backend_port: Option<u16>,
-    request_defaults: Option<serde_json::Map<String, serde_json::Value>>,
-    restored: bool,
-}
+mod options;
 
-fn resolve_serve_listen_host(args: &ServeArgs) -> String {
-    args.host
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string)
-        .unwrap_or_else(|| {
-            if args.lan {
-                "0.0.0.0".to_string()
-            } else {
-                "127.0.0.1".to_string()
-            }
-        })
-}
-
-fn resolve_serve_restore_model(args: &ServeArgs) -> Option<ServeModelRequest> {
-    if args.no_restore_model {
-        return None;
-    }
-    if args
-        .model
-        .as_deref()
-        .is_some_and(|value| !value.trim().is_empty())
-    {
-        return None;
-    }
-    if args.mmproj.is_some() || args.ctx_size.is_some() {
-        return None;
-    }
-    if should_run_server_tui(args) {
-        return None;
-    }
-    let selected = local_state::load_state().ok()?.selected_model?;
-    if selected.model.trim().is_empty() {
-        return None;
-    }
-    Some(ServeModelRequest {
-        model: selected.model,
-        mmproj: selected.mmproj,
-        ctx_size: selected.ctx_size,
-        backend_port: args.backend_port,
-        request_defaults: Some(selected.request_defaults),
-        restored: true,
-    })
-}
-
-fn wait_for_foreground_service(
-    mut rust_gateway: std::process::Child,
-    cloudflared_child: Option<std::process::Child>,
-    port: u16,
-    run_id: &str,
-) -> Result<std::process::ExitStatus> {
-    let status = rust_gateway.wait()?;
-    if let Some(mut tunnel) = cloudflared_child {
-        let _ = tunnel.kill();
-        let _ = tunnel.wait();
-    }
-    let _ = serve_state::remove_serve_pid_info_if_run_id(port, run_id);
-    Ok(status)
-}
-
-fn reject_embedded_serve_backend(args: &ServeArgs) -> Result<()> {
-    let Some(backend_id) = resolve_serve_start_backend(args)? else {
-        return Ok(());
-    };
-    let registry = backend_registry::BackendRegistry::load_current();
-    let Some(backend) = registry.get(&backend_id) else {
-        return Ok(());
-    };
-    if backend.runtime_mode == "embedded" {
-        anyhow::bail!(
-            "{} is an embedded backend. Python control-plane fallback has been removed; use an external-server backend or a backend adapter service.",
-            backend.id
-        );
-    }
-    Ok(())
-}
-
-fn resolve_serve_start_backend(args: &ServeArgs) -> Result<Option<String>> {
-    if let Some(backend) = args
-        .backend
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-    {
-        return Ok(Some(backend.to_string()));
-    }
-    if let Some(default_backend) = args
-        .default_backend
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-    {
-        return Ok(Some(default_backend.to_string()));
-    }
-    if let Some(selected_backend) = local_state::load_state()
-        .ok()
-        .and_then(|state| state.selected_backend)
-        .filter(|value| !value.trim().is_empty())
-    {
-        return Ok(Some(selected_backend));
-    }
-    Ok(backend_registry::BackendRegistry::load_current()
-        .api_payload(backend_registry::BackendScope::Installed)
-        .get("recommended")
-        .and_then(serde_json::Value::as_str)
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string))
-}
-
-fn validate_serve_remote_access_args(args: &ServeArgs) -> Result<()> {
-    if args.cloudflare && !is_loopback_host(args.host.as_deref().unwrap_or("127.0.0.1")) {
-        anyhow::bail!(
-            "--cloudflare keeps OmniInfer on 127.0.0.1; do not combine it with a non-loopback --host."
-        );
-    }
-    if args.cloudflare && args.allow_insecure_lan {
-        anyhow::bail!(
-            "--cloudflare requires an API key and cannot be combined with --allow-insecure-lan."
-        );
-    }
-    if args.cloudflare && args.allow_remote_management {
-        anyhow::bail!(
-            "--cloudflare keeps /omni/* management endpoints local-only; do not use --allow-remote-management."
-        );
-    }
-    if args.cloudflare && args.behind_proxy {
-        anyhow::bail!(
-            "--cloudflare already configures proxy headers; do not combine it with --behind-proxy."
-        );
-    }
-    Ok(())
-}
-
-fn is_loopback_host(host: &str) -> bool {
-    matches!(host.trim(), "" | "127.0.0.1" | "localhost" | "::1")
-}
-
-fn resolve_serve_api_key(args: &ServeArgs, generate_session_key: bool) -> Result<Option<String>> {
-    if generate_session_key {
-        let config = config::load_app_config().unwrap_or_default();
-        let configured = if let Some(value) = args
-            .api_key
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())
-        {
-            Some(value.to_string())
-        } else if !config.api_key.trim().is_empty() {
-            Some(config.api_key)
-        } else if let Ok(value) = env::var("OMNIINFER_API_KEY") {
-            Some(value.trim().to_string()).filter(|value| !value.is_empty())
-        } else {
-            None
-        };
-
-        if let Some(value) = configured {
-            return if value.trim().eq_ignore_ascii_case("auto") {
-                Ok(Some(generate_session_api_key()?))
-            } else {
-                Ok(Some(value))
-            };
-        }
-        return Ok(Some(generate_session_api_key()?));
-    }
-    Ok(args
-        .api_key
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string))
-}
-
-fn resolve_serve_admin_api_key(args: &ServeArgs) -> Result<Option<String>> {
-    let value = args
-        .admin_api_key
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string)
-        .or_else(|| {
-            env::var("OMNIINFER_ADMIN_API_KEY")
-                .ok()
-                .map(|value| value.trim().to_string())
-                .filter(|value| !value.is_empty())
-        });
-    match value.as_deref() {
-        Some(value) if value.eq_ignore_ascii_case("auto") => Ok(Some(generate_session_api_key()?)),
-        _ => Ok(value),
-    }
-}
-
-fn resolve_serve_admin_api_keys(args: &ServeArgs) -> Result<Vec<gateway_auth::GatewayAdminApiKey>> {
-    let raw = args
-        .admin_api_keys
-        .clone()
-        .or_else(|| env::var("OMNIINFER_ADMIN_API_KEYS").ok());
-    parse_admin_api_keys(raw.as_deref())
-}
-
-pub(crate) fn parse_admin_api_keys(
-    raw: Option<&str>,
-) -> Result<Vec<gateway_auth::GatewayAdminApiKey>> {
-    let Some(raw) = raw.map(str::trim).filter(|value| !value.is_empty()) else {
-        return Ok(Vec::new());
-    };
-    let mut entries = Vec::new();
-    for item in raw
-        .split(',')
-        .map(str::trim)
-        .filter(|item| !item.is_empty())
-    {
-        let Some((id, key)) = item.split_once(':').or_else(|| item.split_once('=')) else {
-            anyhow::bail!("invalid admin API key entry '{item}'; expected ID:KEY or ID=KEY");
-        };
-        let id = id.trim();
-        let key = key.trim();
-        if id.is_empty() || key.is_empty() {
-            anyhow::bail!(
-                "invalid admin API key entry '{item}'; admin id and key must be non-empty"
-            );
-        }
-        entries.push(gateway_auth::GatewayAdminApiKey {
-            id: id.to_string(),
-            key: key.to_string(),
-        });
-    }
-    Ok(entries)
-}
-
-fn admin_keys_file_has_entries() -> bool {
-    let path = paths::admin_keys_file();
-    let Ok(raw) = std::fs::read_to_string(path) else {
-        return false;
-    };
-    let Ok(value) = serde_json::from_str::<serde_json::Value>(&raw) else {
-        return false;
-    };
-    let source = value.get("keys").unwrap_or(&value);
-    match source {
-        serde_json::Value::Object(map) => map.values().any(|key| {
-            key.as_str()
-                .map(str::trim)
-                .is_some_and(|key| !key.is_empty())
-        }),
-        serde_json::Value::Array(items) => items.iter().any(|item| {
-            item.get("key")
-                .and_then(serde_json::Value::as_str)
-                .map(str::trim)
-                .is_some_and(|key| !key.is_empty())
-        }),
-        _ => false,
-    }
-}
-
-fn lan_base_urls(config: &config::AppConfig, lan_enabled: bool) -> Vec<String> {
-    if !lan_enabled {
-        return Vec::new();
-    }
-    let host = config.host.trim();
-    if !host.is_empty() && !matches!(host, "0.0.0.0" | "::") {
-        return vec![format!(
-            "http://{}:{}/v1",
-            host.trim_matches(['[', ']']),
-            config.port
-        )];
-    }
-    detect_primary_lan_ipv4()
-        .into_iter()
-        .map(|ip| format!("http://{ip}:{}/v1", config.port))
-        .collect()
-}
-
-fn detect_primary_lan_ipv4() -> Option<String> {
-    let socket = UdpSocket::bind("0.0.0.0:0").ok()?;
-    socket.connect("8.8.8.8:80").ok()?;
-    let addr = socket.local_addr().ok()?;
-    let ip = addr.ip();
-    if ip.is_loopback() || !ip.is_ipv4() {
-        return None;
-    }
-    Some(ip.to_string())
-}
-
-fn generate_session_api_key() -> Result<String> {
-    let token: String = rand::rng()
-        .sample_iter(Alphanumeric)
-        .take(32)
-        .map(char::from)
-        .collect();
-    Ok(format!("oi_{token}"))
-}
-
+pub(crate) use options::parse_admin_api_keys;
+use options::*;
 fn start_rust_gateway_child(
     public_config: &config::AppConfig,
     args: &ServeArgs,
@@ -1285,269 +509,13 @@ pub(crate) fn expand_home_path(value: &str) -> PathBuf {
     path
 }
 
-fn get_serve_health_state(config: &config::AppConfig) -> Result<serde_json::Value> {
-    let url = format!("{}/health?deep=true", config.service_base_url());
-    let response = http_client::get_json(&url, Duration::from_secs(10))?;
-    if response.status >= 400 {
-        anyhow::bail!(
-            "GET /health?deep=true failed with status {}",
-            response.status
-        );
-    }
-    Ok(response.body.get("omni").cloned().unwrap_or(response.body))
-}
+mod health;
 
-fn print_serve_ready(
-    port: u16,
-    state: &serde_json::Value,
-    public_url: Option<&str>,
-    lan_base_urls: &[String],
-    api_key: Option<&str>,
-    admin_api_key: Option<&str>,
-    public_model_root: Option<&std::path::Path>,
-    remote_management: bool,
-    print_api_key: bool,
-    persistent: bool,
-    log_path: &std::path::Path,
-    smoke_text: Option<&str>,
-) {
-    println!();
-    println!("OmniInfer service is ready");
-    let openai_compatible = json_bool(state, "openai_compatible").unwrap_or(true);
-    if let Some(public_url) = public_url {
-        if openai_compatible {
-            println!("OpenAI Base URL: {}/v1", public_url.trim_end_matches('/'));
-        } else {
-            println!("Public Gateway URL: {}", public_url.trim_end_matches('/'));
-        }
-        println!("Health URL: {}/health", public_url.trim_end_matches('/'));
-    }
-    for lan_base_url in lan_base_urls {
-        if openai_compatible {
-            println!("LAN Base URL: {lan_base_url}");
-        } else {
-            println!("LAN Gateway URL: {}", lan_base_url.trim_end_matches("/v1"));
-        }
-    }
-    if openai_compatible {
-        println!("Local Base URL: http://127.0.0.1:{port}/v1");
-    } else {
-        println!("Local Gateway URL: http://127.0.0.1:{port}");
-        if let Some(endpoint) = json_str(state, "client_endpoint") {
-            println!("VLA Client Endpoint: {endpoint}");
-        }
-    }
-    if let Some(api_key) = api_key.filter(|_| print_api_key) {
-        println!("API Key: {api_key}");
-    }
-    if remote_management {
-        println!("Remote management: enabled");
-        if let Some(admin_api_key) = admin_api_key.filter(|_| print_api_key) {
-            println!("Admin API Key: {admin_api_key}");
-        }
-        if let Some(public_model_root) = public_model_root {
-            println!("Public model root: {}", public_model_root.display());
-            match omniinfer_core::public_models::list_public_models(Some(public_model_root)) {
-                Ok(models) => println!("Public models: {}", models.len()),
-                Err(error) => println!("Public models: unavailable ({error})"),
-            }
-        }
-    }
-    println!("Backend: {}", json_str(state, "backend").unwrap_or("-"));
-    if let Some(protocol) = json_str(state, "external_server_protocol") {
-        println!("Backend protocol: {protocol}");
-    }
-    println!(
-        "Backend ready: {}",
-        yes_no(json_bool(state, "backend_ready").unwrap_or(false))
-    );
-    println!("Model: {}", json_str(state, "model").unwrap_or("-"));
-    println!("mmproj: {}", json_str(state, "mmproj").unwrap_or("-"));
-    println!(
-        "ctx-size: {}",
-        json_u64(state, "ctx_size")
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "-".to_string())
-    );
-    if let Some(smoke_text) = smoke_text {
-        println!("Smoke: {smoke_text}");
-    }
-    println!("Log: {}", log_path.display());
-    if !persistent {
-        return;
-    }
-    println!("Stop: ./omniinfer serve stop --port {port}");
-    let remote_base_url = if openai_compatible {
-        public_url
-            .map(|url| format!("{}/v1", url.trim_end_matches('/')))
-            .or_else(|| lan_base_urls.first().cloned())
-    } else {
-        None
-    };
-    if let Some(remote_base_url) = remote_base_url {
-        println!("Curl:");
-        let auth = if let Some(api_key) = api_key.filter(|_| print_api_key) {
-            format!(" -H 'Authorization: Bearer {api_key}'")
-        } else {
-            String::new()
-        };
-        println!(
-            "  curl -sS{} -H 'Content-Type: application/json' {}/chat/completions -d '{{\"model\":\"omniinfer\",\"messages\":[{{\"role\":\"user\",\"content\":\"Hello\"}}],\"stream\":false}}'",
-            auth, remote_base_url
-        );
-        if remote_management {
-            let management_base = remote_base_url.trim_end_matches("/v1");
-            let admin_auth = if let Some(admin_api_key) = admin_api_key.filter(|_| print_api_key) {
-                format!(" -H 'Authorization: Bearer {admin_api_key}'")
-            } else {
-                String::new()
-            };
-            println!(
-                "  curl -sS{} {}/omni/public-models",
-                admin_auth, management_base
-            );
-            println!(
-                "  curl -sS{} -H 'Content-Type: application/json' {}/omni/model/select -d '{{\"model\":\"qwen3.5-4b-q4_k_m\"}}'",
-                admin_auth, management_base
-            );
-        }
-    }
-}
-
-fn serve_smoke(base_url: &str, api_key: Option<&str>) -> Result<String> {
-    let mut payload = serde_json::Map::new();
-    payload.insert("model".to_string(), serde_json::json!("omniinfer"));
-    payload.insert(
-        "messages".to_string(),
-        serde_json::json!([{ "role": "user", "content": "Hello" }]),
-    );
-    payload.insert("temperature".to_string(), serde_json::json!(0));
-    payload.insert("max_tokens".to_string(), serde_json::json!(16));
-    payload.insert("stream".to_string(), serde_json::json!(false));
-
-    let url = format!("{}/v1/chat/completions", base_url.trim_end_matches('/'));
-    let response = http_client::post_json_with_bearer(
-        &url,
-        &serde_json::Value::Object(payload),
-        api_key,
-        Duration::from_secs(120),
-    )?;
-    if response.status >= 400 {
-        anyhow::bail!(
-            "POST /v1/chat/completions failed with status {}",
-            response.status
-        );
-    }
-    response
-        .body
-        .get("choices")
-        .and_then(serde_json::Value::as_array)
-        .and_then(|choices| choices.first())
-        .and_then(|choice| choice.get("message"))
-        .and_then(|message| message.get("content"))
-        .and_then(serde_json::Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_string)
-        .ok_or_else(|| anyhow::anyhow!("Smoke test returned an empty response."))
-}
-
-fn serve_smoke_with_retry(base_url: &str, api_key: Option<&str>) -> Result<String> {
-    let deadline = Instant::now() + public_smoke_retry_duration();
-    loop {
-        match serve_smoke(base_url, api_key) {
-            Ok(text) => return Ok(text),
-            Err(error) => {
-                if !is_transient_public_smoke_error(&error) || Instant::now() >= deadline {
-                    return Err(error);
-                }
-                std::thread::sleep(Duration::from_secs(2));
-            }
-        }
-    }
-}
-
-fn public_smoke_retry_duration() -> Duration {
-    env::var("OMNIINFER_RUST_PUBLIC_SMOKE_RETRY_SECONDS")
-        .ok()
-        .and_then(|value| value.trim().parse::<u64>().ok())
-        .map(Duration::from_secs)
-        .unwrap_or_else(|| Duration::from_secs(45))
-}
-
-fn is_transient_public_smoke_error(error: &anyhow::Error) -> bool {
-    let text = error.to_string().to_ascii_lowercase();
-    text.contains("failed to lookup address")
-        || text.contains("name or service not known")
-        || text.contains("unknown host")
-        || text.contains("no such host")
-        || text.contains("os error 11001")
-        || text.contains("temporary failure in name resolution")
-        || text.contains("connection refused")
-        || text.contains("connection reset")
-        || text.contains("timed out")
-        || text.contains("operation timed out")
-        || text.contains("http status: 520")
-        || text.contains("http status: 521")
-        || text.contains("http status: 522")
-        || text.contains("http status: 523")
-        || text.contains("http status: 524")
-        || text.contains("http status: 530")
-}
-
-pub(crate) fn print_serve_status(port: u16) {
-    let mut config = config::load_app_config().unwrap_or_default();
-    config.port = port;
-    println!("OmniInfer Serve Status");
-    println!("Port: {port}");
-    match serve_state::load_serve_pid_info(port) {
-        Ok(Some(info)) => {
-            if let Some(pid) = info.pid {
-                println!("PID: {pid}");
-            }
-            if let Some(public_url) = info
-                .public_url
-                .as_deref()
-                .filter(|value| !value.trim().is_empty())
-            {
-                println!("OpenAI Base URL: {}/v1", public_url.trim_end_matches('/'));
-            }
-            if let Some(log) = info.log.as_deref().filter(|value| !value.trim().is_empty()) {
-                println!("Log: {log}");
-            }
-        }
-        Ok(None) => {}
-        Err(error) => println!("Serve metadata: unavailable ({error})"),
-    }
-
-    let url = format!("{}/health?deep=true", config.service_base_url());
-    match http_client::get_json(&url, Duration::from_secs(2)) {
-        Ok(response) if response.status == 200 => {
-            let state = response.body.get("omni").unwrap_or(&response.body);
-            println!("Backend: {}", json_str(state, "backend").unwrap_or("-"));
-            if let Some(protocol) = json_str(state, "external_server_protocol") {
-                println!("Backend protocol: {protocol}");
-            }
-            if let Some(endpoint) = json_str(state, "client_endpoint") {
-                println!("Client endpoint: {endpoint}");
-            }
-            println!(
-                "Backend ready: {}",
-                yes_no(json_bool(state, "backend_ready").unwrap_or(false))
-            );
-            println!("Model: {}", json_str(state, "model").unwrap_or("-"));
-            println!(
-                "ctx-size: {}",
-                json_u64(state, "ctx_size")
-                    .map(|value| value.to_string())
-                    .unwrap_or_else(|| "-".to_string())
-            );
-        }
-        Ok(response) => println!("Service: unhealthy (HTTP {})", response.status),
-        Err(error) => println!("OmniInfer service is not running on port {port}: {error}"),
-    }
-}
-
+pub(crate) use health::print_serve_status;
+use health::{
+    get_serve_health_state, is_transient_public_smoke_error, print_serve_ready, serve_smoke,
+    serve_smoke_with_retry,
+};
 fn env_flag(name: &str) -> bool {
     env::var(name)
         .map(|value| {
